Support all Gameevents
We sould do It now like Gmod does it.

+++ source/lua_utils.h
@@ -129,7 +129,7 @@ struct ILuaValue
 {
 	int type = -1;
 
-	int number = -1;
+	double number = -1;
 	const char* string = "";
 	Vector vec;
 	QAngle ang;
